requirements.txt
fastapi
uvicorn
torch
torchvision
Pillow
nest_asyncio
Pillow
wandb
scikit-learn
python-multipart